requirements.txt
numpy==1.21.5
opencv_python_headless==4.6.0.66
Pillow==9.2.0
streamlit==1.12.0
tensorflow==2.9.1
